Version 1.0! Basic features all implemented.
The client can now check for the most recent file in the .dsevents directory, or choose its own file to read if provided with a filename. The output file also now reflects the file read, not the file made as output to make it easier to find which output relates to which .dsevents file. All code has been thoroughly documented.